- Remastered music.  - Added a safeguard in the copper list access.

+++ copper_routines.h
@@ -4,5 +4,5 @@
 */
 
 void DeleteCopperList(void);
-void CreateHigheSTColorCopperList(int scanline_offset, int y_offset);
+void CreateHigheSTColorCopperList(short scanline_offset, short y_offset);
 void CreateVerticalCopperList(short y_offset, /* UWORD base_color,*/ UWORD *palette_array, short palette_size);
